+++ app/src/main/AndroidManifest.xml
@@ -16,8 +16,32 @@
         android:theme="@style/Theme.Trace"
         tools:targetApi="31">
 
+        <activity
+            android:name="com.example.main.MainActivity"
+            android:theme="@style/Theme.Trace.Splash"
+            android:windowSoftInputMode="adjustResize"
+            android:configChanges="uiMode"
+            android:exported="true">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+        </activity>
+
+        <activity
+            android:name="com.kakao.sdk.auth.AuthCodeHandlerActivity"
+            android:exported="true">
+            <intent-filter>
+                <action android:name="android.intent.action.VIEW" />
+                <category android:name="android.intent.category.DEFAULT" />
+                <category android:name="android.intent.category.BROWSABLE" />
 
 
+                <data android:host="oauth"
+                    android:scheme="kakao${KAKAO_NATIVE_APP_KEY}" />
+            </intent-filter>
+        </activity>
 
     </application>
