+++ app/src/main/res/values/themes.xml
@@ -1,4 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
     <style name="Theme.Trace" parent="android:Theme.Material.Light.NoActionBar" />
+
+    <style name="Theme.Trace.Splash" parent="Theme.SplashScreen">
+        <item name="windowSplashScreenBackground">@color/white</item>
+        <item name="postSplashScreenTheme">@style/Theme.Trace</item>
+    </style>
 </resources>
