+++ feature/main/src/main/java/com/example/main/MainViewModel.kt
@@ -6,6 +6,10 @@ import com.example.common.event.EventHelper
 import com.example.domain.repository.UserRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.channels.Channel
+import kotlinx.coroutines.delay
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.receiveAsFlow
 import kotlinx.coroutines.launch
 import javax.inject.Inject
@@ -18,17 +22,23 @@ class MainViewModel @Inject constructor(
     private val _eventChannel = Channel<MainEvent>()
     val eventChannel = _eventChannel.receiveAsFlow()
 
+    private val _isAppReady = MutableStateFlow(false)
+    val isAppReady: StateFlow<Boolean> = _isAppReady.asStateFlow()
+
     internal fun checkSession() = viewModelScope.launch {
         when {
             userRepository.checkSession() -> { _eventChannel.send(MainEvent.NavigateHome) }
-            else -> { _eventChannel.send(MainEvent.NavigateLogin) }
+            else -> { _eventChannel.send(MainEvent.NavigateHome) }
         }
 
     }
 
+    fun onAppReady() {
+        _isAppReady.value = true
+    }
+
     sealed class MainEvent {
         data object NavigateHome : MainEvent()
-        data object NavigateLogin : MainEvent()
     }
 
 }
